application/backend/requirements.txt
alembic==1.13.3
fastapi==0.115.2
psycopg2-binary==2.9.10
pydantic-settings==2.6.0
python-multipart==0.0.12
SQLAlchemy==2.0.36
sqlmodel==0.0.22
uvicorn==0.32.0
pytest==8.3.5
httpx==0.28.1
ruff==0.3.7
mypy==1.9.0
fluent-logger==0.10.0
prometheus-client==0.16.0
sentry-sdk==1.14.0
structlog==22.3.0
